finish python interface
