+++ requirements.txt
@@ -1,0 +1,3 @@
+numpy==1.26
+pandas==2.1
+pybind11==2.13
