+++ src/pybind_OpenLPT/pyOpenLPT.cpp
@@ -1,0 +1,5 @@
+
+void init_OpenLPT(py::module &m) 
+{
+    m.def("run", &run, "run OpenLPT");
+}
