+++ test/inputs/test_STB/imgFile/cam3ImageNames_python.txt
@@ -1,0 +1,50 @@
+../../test/inputs/test_STB/imgFile/cam3/img00000.tif
+../../test/inputs/test_STB/imgFile/cam3/img00001.tif
+../../test/inputs/test_STB/imgFile/cam3/img00002.tif
+../../test/inputs/test_STB/imgFile/cam3/img00003.tif
+../../test/inputs/test_STB/imgFile/cam3/img00004.tif
+../../test/inputs/test_STB/imgFile/cam3/img00005.tif
+../../test/inputs/test_STB/imgFile/cam3/img00006.tif
+../../test/inputs/test_STB/imgFile/cam3/img00007.tif
+../../test/inputs/test_STB/imgFile/cam3/img00008.tif
+../../test/inputs/test_STB/imgFile/cam3/img00009.tif
+../../test/inputs/test_STB/imgFile/cam3/img00010.tif
+../../test/inputs/test_STB/imgFile/cam3/img00011.tif
+../../test/inputs/test_STB/imgFile/cam3/img00012.tif
+../../test/inputs/test_STB/imgFile/cam3/img00013.tif
+../../test/inputs/test_STB/imgFile/cam3/img00014.tif
+../../test/inputs/test_STB/imgFile/cam3/img00015.tif
+../../test/inputs/test_STB/imgFile/cam3/img00016.tif
+../../test/inputs/test_STB/imgFile/cam3/img00017.tif
+../../test/inputs/test_STB/imgFile/cam3/img00018.tif
+../../test/inputs/test_STB/imgFile/cam3/img00019.tif
+../../test/inputs/test_STB/imgFile/cam3/img00020.tif
+../../test/inputs/test_STB/imgFile/cam3/img00021.tif
+../../test/inputs/test_STB/imgFile/cam3/img00022.tif
+../../test/inputs/test_STB/imgFile/cam3/img00023.tif
+../../test/inputs/test_STB/imgFile/cam3/img00024.tif
+../../test/inputs/test_STB/imgFile/cam3/img00025.tif
+../../test/inputs/test_STB/imgFile/cam3/img00026.tif
+../../test/inputs/test_STB/imgFile/cam3/img00027.tif
+../../test/inputs/test_STB/imgFile/cam3/img00028.tif
+../../test/inputs/test_STB/imgFile/cam3/img00029.tif
+../../test/inputs/test_STB/imgFile/cam3/img00030.tif
+../../test/inputs/test_STB/imgFile/cam3/img00031.tif
+../../test/inputs/test_STB/imgFile/cam3/img00032.tif
+../../test/inputs/test_STB/imgFile/cam3/img00033.tif
+../../test/inputs/test_STB/imgFile/cam3/img00034.tif
+../../test/inputs/test_STB/imgFile/cam3/img00035.tif
+../../test/inputs/test_STB/imgFile/cam3/img00036.tif
+../../test/inputs/test_STB/imgFile/cam3/img00037.tif
+../../test/inputs/test_STB/imgFile/cam3/img00038.tif
+../../test/inputs/test_STB/imgFile/cam3/img00039.tif
+../../test/inputs/test_STB/imgFile/cam3/img00040.tif
+../../test/inputs/test_STB/imgFile/cam3/img00041.tif
+../../test/inputs/test_STB/imgFile/cam3/img00042.tif
+../../test/inputs/test_STB/imgFile/cam3/img00043.tif
+../../test/inputs/test_STB/imgFile/cam3/img00044.tif
+../../test/inputs/test_STB/imgFile/cam3/img00045.tif
+../../test/inputs/test_STB/imgFile/cam3/img00046.tif
+../../test/inputs/test_STB/imgFile/cam3/img00047.tif
+../../test/inputs/test_STB/imgFile/cam3/img00048.tif
+../../test/inputs/test_STB/imgFile/cam3/img00049.tif
