+++ test/inputs/test_STB/imgFile/cam4ImageNames_python.txt
@@ -1,0 +1,50 @@
+../../test/inputs/test_STB/imgFile/cam4/img00000.tif
+../../test/inputs/test_STB/imgFile/cam4/img00001.tif
+../../test/inputs/test_STB/imgFile/cam4/img00002.tif
+../../test/inputs/test_STB/imgFile/cam4/img00003.tif
+../../test/inputs/test_STB/imgFile/cam4/img00004.tif
+../../test/inputs/test_STB/imgFile/cam4/img00005.tif
+../../test/inputs/test_STB/imgFile/cam4/img00006.tif
+../../test/inputs/test_STB/imgFile/cam4/img00007.tif
+../../test/inputs/test_STB/imgFile/cam4/img00008.tif
+../../test/inputs/test_STB/imgFile/cam4/img00009.tif
+../../test/inputs/test_STB/imgFile/cam4/img00010.tif
+../../test/inputs/test_STB/imgFile/cam4/img00011.tif
+../../test/inputs/test_STB/imgFile/cam4/img00012.tif
+../../test/inputs/test_STB/imgFile/cam4/img00013.tif
+../../test/inputs/test_STB/imgFile/cam4/img00014.tif
+../../test/inputs/test_STB/imgFile/cam4/img00015.tif
+../../test/inputs/test_STB/imgFile/cam4/img00016.tif
+../../test/inputs/test_STB/imgFile/cam4/img00017.tif
+../../test/inputs/test_STB/imgFile/cam4/img00018.tif
+../../test/inputs/test_STB/imgFile/cam4/img00019.tif
+../../test/inputs/test_STB/imgFile/cam4/img00020.tif
+../../test/inputs/test_STB/imgFile/cam4/img00021.tif
+../../test/inputs/test_STB/imgFile/cam4/img00022.tif
+../../test/inputs/test_STB/imgFile/cam4/img00023.tif
+../../test/inputs/test_STB/imgFile/cam4/img00024.tif
+../../test/inputs/test_STB/imgFile/cam4/img00025.tif
+../../test/inputs/test_STB/imgFile/cam4/img00026.tif
+../../test/inputs/test_STB/imgFile/cam4/img00027.tif
+../../test/inputs/test_STB/imgFile/cam4/img00028.tif
+../../test/inputs/test_STB/imgFile/cam4/img00029.tif
+../../test/inputs/test_STB/imgFile/cam4/img00030.tif
+../../test/inputs/test_STB/imgFile/cam4/img00031.tif
+../../test/inputs/test_STB/imgFile/cam4/img00032.tif
+../../test/inputs/test_STB/imgFile/cam4/img00033.tif
+../../test/inputs/test_STB/imgFile/cam4/img00034.tif
+../../test/inputs/test_STB/imgFile/cam4/img00035.tif
+../../test/inputs/test_STB/imgFile/cam4/img00036.tif
+../../test/inputs/test_STB/imgFile/cam4/img00037.tif
+../../test/inputs/test_STB/imgFile/cam4/img00038.tif
+../../test/inputs/test_STB/imgFile/cam4/img00039.tif
+../../test/inputs/test_STB/imgFile/cam4/img00040.tif
+../../test/inputs/test_STB/imgFile/cam4/img00041.tif
+../../test/inputs/test_STB/imgFile/cam4/img00042.tif
+../../test/inputs/test_STB/imgFile/cam4/img00043.tif
+../../test/inputs/test_STB/imgFile/cam4/img00044.tif
+../../test/inputs/test_STB/imgFile/cam4/img00045.tif
+../../test/inputs/test_STB/imgFile/cam4/img00046.tif
+../../test/inputs/test_STB/imgFile/cam4/img00047.tif
+../../test/inputs/test_STB/imgFile/cam4/img00048.tif
+../../test/inputs/test_STB/imgFile/cam4/img00049.tif
